requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.3
flask-login==0.6.2
psycopg[binary]==3.1.18
python-dotenv==1.0.0
werkzeug==2.3.6